my-api/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
supabase
pandas
redis
python-dotenv
python-multipart
psycopg2-binary
alembic
sqlalchemy
httpx
openai
rich
